refactor(linalg): de-duplicate linalg module (#47)
Fifth de-duplication step of #47, following the established pattern.

- src/linalg/linalg.ts: replace the type-only stub with the REAL
  implementation moved verbatim from the monolith (JacobiImpl, JacobiSVDImpl,
  lu_solve, cholesky_solve, svd_decompose, svd_solve, svd_invert, eigenVV).
  Imports swap/hypot from ./linalg_base and matmath — both already real
  modules.
- src/jsfeatNext.ts: shrinks by ~740 lines (largest single extraction).

Verified behavior-preserving: tsc --noEmit clean; npm test 57/57 (6 parity
tests pin linalg: LU, Cholesky, SVD x3, eigenVV vs original jsfeat); UMD
bundle smoke-checked (instanceof, static inheritance, lu_solve on an SPD
system).
